numbers are not in strings and airity rethink
numbers should not be in strings, so fixed that and Airity of a function is separate from how many aguments a conditional function consumes, so the introduction of the prefix of 2-- onto an function word. For example 2--foo declares that when and if foo is called it will consume two static arguments to its left.